+++ ts/features/bonus/bpd/components/BpdTransactionSummaryComponent.tsx
@@ -1,11 +1,15 @@
 import { View } from "native-base";
 import * as React from "react";
 import { StyleSheet } from "react-native";
+import { TouchableWithoutFeedback } from "@gorhom/bottom-sheet";
 import { Body } from "../../../../components/core/typography/Body";
 import { H4 } from "../../../../components/core/typography/H4";
 import { H5 } from "../../../../components/core/typography/H5";
 import { InfoBox } from "../../../../components/box/InfoBox";
 import { useIOBottomSheet } from "../../../../utils/bottomSheet";
+import { Link } from "../../../../components/core/typography/Link";
+import { openWebUrl } from "../../../../utils/url";
+import Markdown from "../../../../components/ui/Markdown";
 import { IOColors } from "../../../../components/core/variables/IOColors";
 import TouchableDefaultOpacity from "../../../../components/TouchableDefaultOpacity";
 import IconFont from "../../../../components/ui/IconFont";
@@ -17,7 +21,6 @@ import {
 } from "../../../../utils/stringBuilder";
 import { BpdAmount } from "../saga/networking/amount";
 import { BpdPeriod } from "../store/actions/periods";
-import { BottomSheetBpdTransactionsBody } from "./BottomSheetBpdTransactionsBody";
 
 type Props = {
   lastUpdateDate: string;
@@ -29,8 +32,49 @@ const styles = StyleSheet.create({
   row: {
     flexDirection: "row",
     alignItems: "center"
-  }
+  },
+  readMore: { marginLeft: 31, marginBottom: 24 }
 });
+
+const readMoreLink = "https://io.italia.it/cashback/acquirer/";
+const SHOW_CTA_DELAY = 500;
+
+const CSS_STYLE = `
+body {
+  font-size: 16;
+  color: ${IOColors.black}
+}
+`;
+
+export const BottomSheetBpdTransactionsBody: React.FunctionComponent = () => {
+  const [CTAVisibility, setCTAVisibility] = React.useState(false);
+
+  const setCTAVisible = () =>
+    setTimeout(() => setCTAVisibility(true), SHOW_CTA_DELAY);
+
+  return (
+    <>
+      <Markdown
+        cssStyle={CSS_STYLE}
+        avoidTextSelection={true}
+        onLoadEnd={setCTAVisible}
+      >
+        {I18n.t(
+          "bonus.bpd.details.transaction.detail.summary.bottomSheet.body"
+        )}
+      </Markdown>
+      {CTAVisibility && (
+        <TouchableWithoutFeedback onPress={() => openWebUrl(readMoreLink)}>
+          <Link style={styles.readMore} weight={"SemiBold"}>
+            {I18n.t(
+              "bonus.bpd.details.transaction.detail.summary.bottomSheet.readMore"
+            )}
+          </Link>
+        </TouchableWithoutFeedback>
+      )}
+    </>
+  );
+};
 
 const BpdTransactionSummaryComponent: React.FunctionComponent<Props> = (
   props: Props
