+++ ts/features/bonus/bpd/screens/details/transaction/BpdEmptyTransactionsList.tsx
@@ -4,7 +4,7 @@ import { InfoBox } from "../../../../../../components/box/InfoBox";
 import { Body } from "../../../../../../components/core/typography/Body";
 import { H4 } from "../../../../../../components/core/typography/H4";
 import I18n from "../../../../../../i18n";
-import { BottomSheetBpdTransactionsBody } from "../../../components/BottomSheetBpdTransactionsBody";
+import { BottomSheetBpdTransactionsBody } from "../../../components/BpdTransactionSummaryComponent";
 
 const BpdEmptyTransactionsList: React.FunctionComponent = () => (
   <>
